prevent users from changing their own permissions or deleting themselves

+++ android/application/app/src/main/java/de/unihannover/se/tauben2/view/recycler/UsersRecyclerFragment.kt
@@ -1,5 +1,6 @@
 package de.unihannover.se.tauben2.view.recycler
 
+import android.view.View.INVISIBLE
 import androidx.appcompat.app.AlertDialog
 import androidx.databinding.ViewDataBinding
 import de.unihannover.se.tauben2.R
@@ -23,29 +24,36 @@ class UsersRecyclerFragment : RecyclerFragment<User>() {
         this.user = data
         if (binding is CardUserBinding) {
             binding.c = data
-            binding.cardUserIsAuthorized.setOnClickListener {
-                vm?.updatePermissions(data.username, Auth(data.isActivated, data.isAdmin))
-            }
-            binding.cardUserIsAdmin.setOnClickListener {
-                vm?.updatePermissions(data.username, Auth(data.isActivated, data.isAdmin))
-            }
-            binding.cardBtnUserDelete.setOnClickListener {
-                context?.run {
-                    AlertDialog.Builder(this).apply {
-                        setTitle(getString(R.string.title_deleting_user))
-                        setMessage(getString(R.string.desc_delete_user))
-
-                        setPositiveButton(R.string.delete) { _, _ ->
-                            vm?.deleteUser(data)
-                            setSnackBar(binding.root, getString(R.string.alert_user_deleted))
-                        }
-
-                        setNegativeButton(R.string.cancel) { dialogInterface, _ ->
-                            dialogInterface.cancel()
-                        }
-
-                    }.show()
+            if (data.username != vm?.getOwnerUsername()) {
+                binding.cardUserIsAuthorized.setOnClickListener {
+                    vm?.updatePermissions(data.username, Auth(data.isActivated, data.isAdmin))
                 }
+                binding.cardUserIsAdmin.setOnClickListener {
+                    vm?.updatePermissions(data.username, Auth(data.isActivated, data.isAdmin))
+                }
+                binding.cardBtnUserDelete.setOnClickListener {
+                    context?.run {
+                        AlertDialog.Builder(this).apply {
+                            setTitle(getString(R.string.title_deleting_user))
+                            setMessage(getString(R.string.desc_delete_user))
+
+                            setPositiveButton(R.string.delete) { _, _ ->
+                                vm?.deleteUser(data)
+                                setSnackBar(binding.root, getString(R.string.alert_user_deleted))
+                            }
+
+                            setNegativeButton(R.string.cancel) { dialogInterface, _ ->
+                                dialogInterface.cancel()
+                            }
+
+                        }.show()
+                    }
+
+                }
+            } else {
+                binding.cardUserIsAuthorized.isEnabled = false
+                binding.cardUserIsAdmin.isEnabled = false
+                binding.cardBtnUserDelete.visibility = INVISIBLE
 
             }
         }
